fix styling of the tab navbar; split navigators
