+++ App.js
@@ -1,9 +1,7 @@
 import { StatusBar } from 'expo-status-bar';
 import React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
-import BottomTabNavigator from './navigation/BottomTabNavigator';
-// import MainNavigation from './navigation/BottomTabNavigator';
-import DrawerNavigator from './navigation/BottomTabNavigator';
+import DrawerNavigator from './navigation/DrawerNavigation';
 
 export default function App() {
   return (
